Using Python to run ML techniques for an e-commerce company.
An international e-commerce company  based that sells electronic products wants to discover key insights from their customer database. I am using different ML techniques to predict whether the shipment reached on time. The variable that we are trying to predict is Reached on time: It is the target variable, where 1 Indicates that the product has NOT reached on time and 0 indicates it has reached on time.